Change workshop to event for terminology reasons (#62)
This is needed because currently we only have workshops. In the future, we want to have at least two types of events and a workshop is one of them. We need a precise terminology for this, so workshop was changed to event, which will be the generic term for any kind of event.